Upgrade WorkshopManagementEventHandler to .NET 6
