+++ src/WorkshopManagementEventHandler/Events/CustomerRegistered.cs
@@ -1,22 +1,16 @@
-﻿using Pitstop.Infrastructure.Messaging;
-using System;
-using System.Collections.Generic;
-using System.Text;
+﻿namespace Pitstop.WorkshopManagementEventHandler.Events;
 
-namespace Pitstop.WorkshopManagementEventHandler.Events
+public class CustomerRegistered : Event
 {
-    public class CustomerRegistered : Event
-    {
-        public readonly string CustomerId;
-        public readonly string Name;
-        public readonly string TelephoneNumber;
+    public readonly string CustomerId;
+    public readonly string Name;
+    public readonly string TelephoneNumber;
 
-        public CustomerRegistered(Guid messageId, string customerId, string name, string telephoneNumber) : 
-            base(messageId)
-        {
-            CustomerId = customerId;
-            Name = name;
-            TelephoneNumber = telephoneNumber;
-        }
+    public CustomerRegistered(Guid messageId, string customerId, string name, string telephoneNumber) :
+        base(messageId)
+    {
+        CustomerId = customerId;
+        Name = name;
+        TelephoneNumber = telephoneNumber;
     }
 }
