+++ src/WorkshopManagementEventHandler/Events/MaintenanceJobFinished.cs
@@ -1,24 +1,18 @@
-﻿using Pitstop.Infrastructure.Messaging;
-using System;
-using System.Collections.Generic;
-using System.Text;
+﻿namespace Pitstop.WorkshopManagementEventHandler.Events;
 
-namespace Pitstop.WorkshopManagementEventHandler.Events
+public class MaintenanceJobFinished : Event
 {
-    public class MaintenanceJobFinished : Event
-    {
-        public readonly Guid JobId;
-        public readonly DateTime StartTime;
-        public readonly DateTime EndTime;
-        public readonly string Notes;
+    public readonly Guid JobId;
+    public readonly DateTime StartTime;
+    public readonly DateTime EndTime;
+    public readonly string Notes;
 
-        public MaintenanceJobFinished(Guid messageId, Guid jobId, DateTime startTime, DateTime endTime, string notes) : 
-            base(messageId)
-        {
-            JobId = jobId;
-            StartTime = startTime;
-            EndTime = endTime;
-            Notes = notes;
-        }
+    public MaintenanceJobFinished(Guid messageId, Guid jobId, DateTime startTime, DateTime endTime, string notes) :
+        base(messageId)
+    {
+        JobId = jobId;
+        StartTime = startTime;
+        EndTime = endTime;
+        Notes = notes;
     }
 }
