+++ src/WorkshopManagementEventHandler/GlobalUsings.cs
@@ -1,0 +1,13 @@
+global using Pitstop.WorkshopManagementEventHandler;
+global using Pitstop.WorkshopManagementEventHandler.Events;
+global using Pitstop.WorkshopManagementEventHandler.Model;
+global using Pitstop.Infrastructure.Messaging;
+global using Pitstop.Infrastructure.Messaging.Configuration;
+global using Pitstop.WorkshopManagementEventHandler.DataAccess;
+global using Microsoft.EntityFrameworkCore;
+global using Microsoft.Extensions.Configuration;
+global using Microsoft.Extensions.DependencyInjection;
+global using Microsoft.Extensions.Hosting;
+global using Polly;
+global using Serilog;
+global using Newtonsoft.Json.Linq;
