+++ src/WorkshopManagementEventHandler/Model/Customer.cs
@@ -1,12 +1,8 @@
-﻿using Pitstop.WorkshopManagementEventHandler.Model;
-using System.Collections.Generic;
+﻿namespace Pitstop.WorkshopManagementEventHandler;
 
-namespace Pitstop.WorkshopManagementEventHandler
+public class Customer
 {
-    public class Customer
-    {
-        public string CustomerId { get; set; }
-        public string Name { get; set; }
-        public string TelephoneNumber { get; set; }
-    }
+    public string CustomerId { get; set; }
+    public string Name { get; set; }
+    public string TelephoneNumber { get; set; }
 }
