+++ src/WorkshopManagementEventHandler/Model/MaintenanceJob.cs
@@ -1,21 +1,15 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Threading.Tasks;
+﻿namespace Pitstop.WorkshopManagementEventHandler.Model;
 
-namespace Pitstop.WorkshopManagementEventHandler.Model
+public class MaintenanceJob
 {
-    public class MaintenanceJob
-    {
-        public Guid Id { get; set; }
-        public DateTime StartTime { get; set; }
-        public DateTime EndTime { get; set; }
-        public Vehicle Vehicle { get; set; }
-        public Customer Customer { get; set; }
-        public string Description { get; set; }
-        public DateTime? ActualStartTime { get; set; }
-        public DateTime? ActualEndTime { get; set; }
-        public string Notes { get; set; }
-        public DateTime WorkshopPlanningDate { get; set; }
-    }
+    public Guid Id { get; set; }
+    public DateTime StartTime { get; set; }
+    public DateTime EndTime { get; set; }
+    public Vehicle Vehicle { get; set; }
+    public Customer Customer { get; set; }
+    public string Description { get; set; }
+    public DateTime? ActualStartTime { get; set; }
+    public DateTime? ActualEndTime { get; set; }
+    public string Notes { get; set; }
+    public DateTime WorkshopPlanningDate { get; set; }
 }
