+++ src/WorkshopManagementEventHandler/Model/Vehicle.cs
@@ -1,10 +1,9 @@
-﻿namespace Pitstop.WorkshopManagementEventHandler.Model
+﻿namespace Pitstop.WorkshopManagementEventHandler.Model;
+
+public class Vehicle
 {
-    public class Vehicle
-    {
-        public string LicenseNumber { get; set; }
-        public string Brand { get; set; }
-        public string Type { get; set; }
-        public string OwnerId { get; set; }
-    }
+    public string LicenseNumber { get; set; }
+    public string Brand { get; set; }
+    public string Type { get; set; }
+    public string OwnerId { get; set; }
 }
