+++ src/WorkshopManagementEventHandler/Program.cs
@@ -1,69 +1,29 @@
-﻿using Microsoft.EntityFrameworkCore;
-using Microsoft.Extensions.Configuration;
-using Microsoft.Extensions.DependencyInjection;
-using Microsoft.Extensions.Hosting;
-using Pitstop.Infrastructure.Messaging;
-using Pitstop.Infrastructure.Messaging.Configuration;
-using Pitstop.WorkshopManagementEventHandler.DataAccess;
-using Polly;
-using Serilog;
-using System;
-using System.IO;
-using System.Threading;
-using System.Threading.Tasks;
-
-namespace Pitstop.WorkshopManagementEventHandler
-{
-    class Program
+﻿IHost host = Host
+    .CreateDefaultBuilder(args)
+    .ConfigureServices((hostContext, services) =>
     {
-        public static async Task Main(string[] args)
+        services.UseRabbitMQMessageHandler(hostContext.Configuration);
+
+        services.AddTransient<WorkshopManagementDBContext>((svc) =>
         {
-            var host = CreateHostBuilder(args).Build();
-            await host.RunAsync();
-        }
+            var sqlConnectionString = hostContext.Configuration.GetConnectionString("WorkshopManagementCN");
+            var dbContextOptions = new DbContextOptionsBuilder<WorkshopManagementDBContext>()
+                .UseSqlServer(sqlConnectionString)
+                .Options;
+            var dbContext = new WorkshopManagementDBContext(dbContextOptions);
 
-        private static IHostBuilder CreateHostBuilder(string[] args)
-        {
-            var hostBuilder = Host.CreateDefaultBuilder(args)
-                .ConfigureHostConfiguration(configHost =>
-                {
-                    configHost.SetBasePath(Directory.GetCurrentDirectory());
-                    configHost.AddJsonFile("hostsettings.json", optional: true);
-                    configHost.AddJsonFile($"appsettings.json", optional: false);
-                    configHost.AddEnvironmentVariables();
-                    configHost.AddEnvironmentVariables("DOTNET_");
-                    configHost.AddCommandLine(args);
-                })
-                .ConfigureAppConfiguration((hostContext, config) =>
-                {
-                    config.AddJsonFile($"appsettings.{hostContext.HostingEnvironment.EnvironmentName}.json", optional: false);
-                })
-                .ConfigureServices((hostContext, services) =>
-                {
-                    services.UseRabbitMQMessageHandler(hostContext.Configuration);
+            DBInitializer.Initialize(dbContext);
 
-                    services.AddTransient<WorkshopManagementDBContext>((svc) =>
-                    {
-                        var sqlConnectionString = hostContext.Configuration.GetConnectionString("WorkshopManagementCN");
-                        var dbContextOptions = new DbContextOptionsBuilder<WorkshopManagementDBContext>()
-                            .UseSqlServer(sqlConnectionString)
-                            .Options;
-                        var dbContext = new WorkshopManagementDBContext(dbContextOptions);
+            return dbContext;
+        });
 
-                        DBInitializer.Initialize(dbContext);
+        services.AddHostedService<EventHandlerWorker>();
+    })
+    .UseSerilog((hostContext, loggerConfiguration) =>
+    {
+        loggerConfiguration.ReadFrom.Configuration(hostContext.Configuration);
+    })
+    .UseConsoleLifetime()
+    .Build();
 
-                        return dbContext;
-                    });
-
-                    services.AddHostedService<EventHandler>();
-                })
-                .UseSerilog((hostContext, loggerConfiguration) =>
-                {
-                    loggerConfiguration.ReadFrom.Configuration(hostContext.Configuration);
-                })
-                .UseConsoleLifetime();
-
-            return hostBuilder;
-        }
-    }
-}
+await host.RunAsync();
